requirements.txt
Pillow
feedparser
requests
loguru